Pop Up Class
This class makes a new JPanel pop up. It is designed to take in a parameter of a JScrollpanel, which ideally holds a JTable of data. It also takes in a string, which is meant to have the title of the table. Then it makes a new pop up that shows a table, and title for the table data.